Add location tooltip for sewer chart
